requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
torch>=1.9.0
torchvision>=0.10.0
pytesseract>=0.3.8
easyocr>=1.4.2
paddleocr>=2.6.0
ultralytics>=8.0.0  # YOLO
scikit-image>=0.18.3
pandas>=1.3.0
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3
PyYAML>=6.0
pytest>=6.2.5 